decouple test execution from Reconcile call count
the test as written previously required knowing the exact number of
times that the Reconcile function would be called. this was problimatic
since things like the Deployment controller would update the Deployment
in ways that were out of our control.

by buffering the request channel we are able to ignore the exact number
of Reconile calls and instead use Eventully to wait for the actual
states we are looking for