stops listening to Deployment/Secret/Service evnts
to prevent unnecasarily releasing new signed metadata we want to reduce
the number of times the Reconcile func is called, to do this we have
stopped listening to the events from changes to the Deployments, Secrets
and Services and now only trigger on changes to the Metadata resources.

this means these resources will no longer "self-heal" if deleted, but
nudging the Metadata resource (by rolling out a new deployment for
example) will still trigger the healing.