EID-1549 Allow Metadata kind to accept certs
We have a use case for the Metadata kind where we want to manually pass
in the signing and encryption certs (the connector node). This is so
that they can be included in metadata signed by the HSM - this will
mean the proxy node and the connector node metadata will have the same
root of trust.

This also updates how the hsmclient interacts with the `mdgen` java app
which actually generates the metadata.