EID-1549 Add `PublishingPath` field to metadata
The eIDAS convention is to publish connector metadata at
`/ConnectorMetadata`. We add a field here to the metadata spec which is
used to configure where the metadata is mounted in our nginx container
in the metadata's deployment.

If the field is not supplied, we default to `metadata.xml` to ensure
backwards compatibility.