EID-1553 Include tags with supplied certs
The `keystore` utility expects certificates with the usual tags,
`-----BEGIN/END CERT..`. Certificates supplied by the HSM include these.

When we specify certificates in our config, for example in the
verify-metadata repo we usually just specify the base64 encoded string
without these tags.

This adds the tags to supplied certificates so we can continue to define
certs in the usual way.