EID-1615 Use timezone and convert in mdgen
We need to store the certificate expiry as a string with a timezone
format in the metadata controller so that when it's converted back to an
object by mdgen the time is correct.

This also updates the integration tests for mdgen to use the new
parameter, as well as to check we don't include the encryption
certificate in the proxy-node use case - this should have been done
before.