EID-1617 Allow connector-node metadata to remount
The way we were mounting the metadata secret volume into the container
was preventing the volume from being automatically remounted if it
changed. Something to do with how `SubPath` used symlinks.

[The docs](https://kubernetes.io/docs/concepts/configuration/secret/#using-secrets)
suggest doing it this way, and this is how we already do it [for the
proxy-node](https://github.com/alphagov/verify-proxy-node/blob/master/chart/templates/stub-connector-deployment.yaml#L35)

Also adds `ReadOnly` because we shouldn't be changing these from within
the container.