BAU: Remove VMC from imported packages
The VMC project should be placed in a specific location to maintain the Go project structure. Then Go doesn't need to clone VMC from GitHub.
Set your GOPATH to a dir and them place the VMC project in
`$GOPATH/src/github.com/alphagov/verify-metadata-controller`

Thanks to @tomrosier for the tip :+1:

Author: @CharlesIC