BAU Move hsm ip resolution out of Dockerfile
The HSM ip address used by the cloudhsm daemon was being baked in to the
Docker image. We've found that the call to dig is now returning more
than one IP address. This change selects the first IP returned whenever
the image is started which should hopefully make it more resilient.