BAU Remove deployment and service from VMC
The VMC should be responsible for generating metadata. It shouldn't be
responsible for also creating a deployment and a service for this
metadata to be accessed via.

Removing this code will require the addition of a Deployment and Service
configuration adding to the existing configuration in the
verify-metadata/ci repo.

This logic isn't needed at all for the proxy-node, as the metadata
secret is mounted directly into the gateway app, and served from there.