EID-1649 Refactor auto refresh logic
This commit changes the way the signing cert is regenerated. There is
now a field on the `CertificateRequestSpec` struct, `RegenerateWhenDaysLeft`.

This field is a pointer to an int (being a pointer it can have a nil
value). This field is passed to the self signed certs config map and can
be used when re-signing the metadata to determine if the signing cert
should be regenerated or just pulled from the config map.

There's also some logic updates to trigger a refresh of the metadata if
the cert needs updating - it's feasible to be in a situation where the
signed metadata doesn't need regenerating, but the saml cert does.

The logic also allows for updating of the `RegenerateWhenDaysLeft` field
without causing the cert itself to be regenerated. This is useful if you
want to either flag a particular cert to be automatically regenerated
when it's expiring, or to do the reverse and stop it being regenerated.

If we don't define the field, we shouldn't regenerate the cert.

This tests also extends the tests to cover the new scenarios.