EID-1941: Upgrade cloudhsm tooling to v3.0.0
We need to update the algorithm we use to sign certificates and
metadata. We have two tools that perform these tasks for us -
`cloudhsmtool` and `mdgen`. Both use HSM to sign.

The java library provided by the AWS cloudhsm tooling includes the
Cavium security provider. Version 2 of Cavium does not appear to support
the RSAPSS algorithms we're required to sign with. Version 3 does.