EID-1941: Get public keys from HSM with Cavium util
The new version of the Cavium library has an interesting new "feature".
The `getKey` method on the CaviumKeyStore now explicitly returns null if
the key being got is a PublicKey. This broke the `cloudhsmtool`.

Fortunately after decompiling and inspecting the library it turns out
that the Cavium libary has a Util class with a `findFirstCaviumKey`
method which will fetch the public keys for us.